backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    MONGODB_URL: str = os.getenv('MONGODB_URL', 'mongodb://localhost:27017')
    DATABASE_NAME: str = os.getenv('DATABASE_NAME', 'finance_tracker')
    SECRET_KEY: str = os.getenv('SECRET_KEY', 'super-secret-key-2024')
    ALGORITHM: str = os.getenv('ALGORITHM', 'HS256')
    ACCESS_TOKEN_EXPIRE_MINUTES: int = int(os.getenv('ACCESS_TOKEN_EXPIRE_MINUTES', '60'))
    OPENAI_API_KEY: str = os.getenv('OPENAI_API_KEY', '')
    ENVIRONMENT: str = os.getenv('ENVIRONMENT', 'development')

settings = Settings()
