backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
bcrypt==4.1.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2
reportlab==4.0.8
openpyxl==3.1.2
matplotlib==3.8.2
